[ElemenBinary] Kernel refactor
