+++ include/flexflow/ops/element_binary.h
@@ -1,33 +1,12 @@
 #ifndef _FLEXFLOW_ELEMENT_BINARY_H
 #define _FLEXFLOW_ELEMENT_BINARY_H
 
-#include "flexflow/device.h"
-#include "flexflow/fftype.h"
 #include "flexflow/layer.h"
 #include "flexflow/node.h"
-#include "flexflow/op_meta.h"
 #include "flexflow/operator.h"
 #include "flexflow/ops/element_binary_params.h"
 
 namespace FlexFlow {
-
-class ElementBinaryMeta : public OpMeta {
-public:
-  ElementBinaryMeta(FFHandler handle);
-#if defined(FF_USE_CUDA) || defined(FF_USE_HIP_CUDA)
-  cudnnTensorDescriptor_t input1Tensor, input2Tensor, outputTensor;
-  cudnnOpTensorDescriptor_t opDesc;
-  cudnnReduceTensorDescriptor_t reduceAddDesc;
-#else
-  miopenTensorDescriptor_t input1Tensor, input2Tensor, outputTensor;
-  miopenTensorOp_t opDesc;
-  miopenReduceTensorDescriptor_t reduceAddDesc;
-#endif
-  OperatorType op_type;
-  bool inplace_a, has_same_operands;
-  bool broadcast_input1, broadcast_input2;
-  char op_name[MAX_OPNAME];
-};
 
 class ElementBinary : public Op {
 public:
@@ -71,32 +50,6 @@ public:
                             std::vector<Legion::PhysicalRegion> const &regions,
                             Legion::Context ctx,
                             Legion::Runtime *runtime);
-  static void init_kernel(ElementBinaryMeta *m,
-                          Legion::Domain const &input1_domain,
-                          Legion::Domain const &input2_domain,
-                          Legion::Domain const &output_domain);
-  static void forward_kernel(ElementBinaryMeta const *m,
-                             float const *in1_ptr,
-                             float const *in2_ptr,
-                             float *out_ptr,
-                             ffStream_t stream);
-  static void forward_kernel_wrapper(ElementBinaryMeta const *m,
-                                     float const *in1_ptr,
-                                     float const *in2_ptr,
-                                     float *out_ptr);
-  static void backward_kernel(ElementBinaryMeta const *m,
-                              float const *out_grad_ptr,
-                              float const *in1_ptr,
-                              float const *in2_ptr,
-                              float *in1_grad_ptr,
-                              float *in2_grad_ptr,
-                              ffStream_t stream);
-  static void backward_kernel_wrapper(ElementBinaryMeta const *m,
-                                      float const *out_grad_ptr,
-                                      float const *in1_ptr,
-                                      float const *in2_ptr,
-                                      float *in1_grad_ptr,
-                                      float *in2_grad_ptr);
   bool measure_operator_cost(Simulator *sim,
                              MachineView const &pc,
                              CostMetrics &cost_metrics) const override;
