+++ src/ops/kernels/element_binary_kernels.cpp
@@ -13,7 +13,7 @@
  * limitations under the License.
  */
 
-#include "flexflow/ops/element_binary.h"
+#include "flexflow/ops/kernels/element_binary_kernels.h"
 #include "flexflow/utils/hip_helper.h"
 #include <hip/hip_runtime.h>
 
@@ -22,11 +22,22 @@ namespace FlexFlow {
 using Legion::coord_t;
 using Legion::Domain;
 
+ElementBinaryMeta::ElementBinaryMeta(FFHandler handler) : OpMeta(handler) {
+  checkCUDNN(miopenCreateTensorDescriptor(&input1Tensor));
+  checkCUDNN(miopenCreateTensorDescriptor(&input2Tensor));
+  checkCUDNN(miopenCreateTensorDescriptor(&outputTensor));
+  checkCUDNN(miopenCreateReduceTensorDescriptor(&reduceAddDesc));
+  op_type = OP_NOOP;
+}
+
+namespace Kernels {
+namespace ElementBinary {
+
 /*static*/
-void ElementBinary::init_kernel(ElementBinaryMeta *m,
-                                Domain const &input1_domain,
-                                Domain const &input2_domain,
-                                Domain const &output_domain) {
+void init_kernel(ElementBinaryMeta *m,
+                 Domain const &input1_domain,
+                 Domain const &input2_domain,
+                 Domain const &output_domain) {
   miopenTensorOp_t mode;
   switch (m->op_type) {
     case OP_EW_ADD:
@@ -53,6 +64,103 @@ void ElementBinary::init_kernel(ElementBinaryMeta *m,
   checkCUDNN(
       cudnnSetTensorDescriptorFromDomain(m->outputTensor, output_domain));
 }
+
+/*static*/
+void forward_kernel_wrapper(ElementBinaryMeta const *m,
+                            float const *in1_ptr,
+                            float const *in2_ptr,
+                            float *out_ptr) {
+  hipStream_t stream;
+  checkCUDA(get_legion_stream(&stream));
+
+  hipEvent_t t_start, t_end;
+  if (m->profiling) {
+    hipEventCreate(&t_start);
+    hipEventCreate(&t_end);
+    hipEventRecord(t_start, stream);
+  }
+  // print_tensor<float>(in1_ptr, in1_domain.get_volume(), "input1:");
+  // print_tensor<float>(in2_ptr, in2_domain.get_volume(), "input2:");
+  Internal::forward_kernel(m, in1_ptr, in2_ptr, out_ptr, stream);
+  // print_tensor<float>(out_ptr, in1_domain.get_volume(), "output:");
+  if (m->profiling) {
+    hipEventRecord(t_end, stream);
+    checkCUDA(hipEventSynchronize(t_end));
+    float elapsed = 0;
+    checkCUDA(hipEventElapsedTime(&elapsed, t_start, t_end));
+    hipEventDestroy(t_start);
+    hipEventDestroy(t_end);
+    char const *opName;
+    switch (m->op_type) {
+      case OP_EW_ADD:
+        opName = "Add";
+        break;
+      case OP_EW_SUB:
+        opName = "Sub";
+        break;
+      case OP_EW_MUL:
+        opName = "Mul";
+        break;
+      case OP_EW_DIV:
+        opName = "Div";
+        break;
+      default:
+        assert(false);
+    }
+    log_measure.debug("[%s] forward time (CF) = %.2fms\n", opName, elapsed);
+  }
+}
+
+/*static*/
+void backward_kernel_wrapper(ElementBinaryMeta const *m,
+                             float const *out_grad_ptr,
+                             float const *in1_ptr,
+                             float const *in2_ptr,
+                             float *in1_grad_ptr,
+                             float *in2_grad_ptr) {
+  hipStream_t stream;
+  checkCUDA(get_legion_stream(&stream));
+  hipEvent_t t_start, t_end;
+  if (m->profiling) {
+    hipEventCreate(&t_start);
+    hipEventCreate(&t_end);
+    hipEventRecord(t_start, stream);
+  }
+
+  Internal::backward_kernel(
+      m, out_grad_ptr, in1_ptr, in2_ptr, in1_grad_ptr, in2_grad_ptr, stream);
+  // elewise_binary_backward_kernel<<<GET_BLOCKS(out_grad_domain.get_volume()),
+  // CUDA_NUM_THREADS>>>( out_grad_domain.get_volume(), alpha, alpha,
+  // ele->op_type, out_grad_ptr, in1_ptr, in2_ptr, in1_grad_ptr, in2_grad_ptr);
+  if (m->profiling) {
+    hipEventRecord(t_end, stream);
+    checkCUDA(hipEventSynchronize(t_end));
+    float elapsed = 0;
+    checkCUDA(hipEventElapsedTime(&elapsed, t_start, t_end));
+    hipEventDestroy(t_start);
+    hipEventDestroy(t_end);
+    char const *opName;
+    switch (m->op_type) {
+      case OP_EW_ADD:
+        opName = "Add";
+        break;
+      case OP_EW_SUB:
+        opName = "Sub";
+        break;
+      case OP_EW_MUL:
+        opName = "Mul";
+        break;
+      case OP_EW_DIV:
+        opName = "Div";
+        break;
+      default:
+        assert(false);
+    }
+    printf("[%s] backward time (CB) = %.2fms\n", opName, elapsed);
+  }
+}
+
+namespace Internal {
 
 __global__ void elewise_binary_forward_kernel(coord_t volume,
                                               float const alpha,
@@ -91,12 +199,50 @@ __global__ void elewise_binary_forward_kernel(coord_t volume,
   }
 }
 
+__global__ void elewise_binary_backward_kernel(coord_t volume,
+                                               float const alpha,
+                                               float const beta,
+                                               OperatorType type,
+                                               float const *out_grad,
+                                               float const *in1,
+                                               float const *in2,
+                                               float *in1_grad,
+                                               float *in2_grad) {
+  CUDA_KERNEL_LOOP(i, volume) {
+    switch (type) {
+      case OP_EW_ADD: {
+        in1_grad[i] = alpha * out_grad[i] + beta * in1_grad[i];
+        in2_grad[i] = alpha * out_grad[i] + beta * in2_grad[i];
+        break;
+      }
+      case OP_EW_SUB: {
+        in1_grad[i] = alpha * out_grad[i] + beta * in1_grad[i];
+        in2_grad[i] = -alpha * out_grad[i] + beta * in2_grad[i];
+        break;
+      }
+      case OP_EW_MUL: {
+        in1_grad[i] = alpha * out_grad[i] * in2[i] + beta * in1_grad[i];
+        in2_grad[i] = alpha * out_grad[i] * in1[i] + beta * in2_grad[i];
+        break;
+      }
+      case OP_EW_DIV: {
+        in1_grad[i] = alpha * out_grad[i] / in2[i] + beta * in1_grad[i];
+        in2_grad[i] = -alpha * out_grad[i] * in1[i] / (in2[i] * in2[i]) +
+                      beta * in2_grad[i];
+        break;
+      }
+      default:
+        assert(false);
+    }
+  }
+}
+
 /*static*/
-void ElementBinary::forward_kernel(ElementBinaryMeta const *m,
-                                   float const *in1_ptr,
-                                   float const *in2_ptr,
-                                   float *out_ptr,
-                                   hipStream_t stream) {
+void forward_kernel(ElementBinaryMeta const *m,
+                    float const *in1_ptr,
+                    float const *in2_ptr,
+                    float *out_ptr,
+                    hipStream_t stream) {
   checkCUDA(hipblasSetStream(m->handle.blas, stream));
   checkCUDNN(miopenSetStream(m->handle.dnn, stream));
 
@@ -154,97 +300,13 @@ void ElementBinary::forward_kernel(ElementBinaryMeta const *m,
 }
 
 /*static*/
-void ElementBinary::forward_kernel_wrapper(ElementBinaryMeta const *m,
-                                           float const *in1_ptr,
-                                           float const *in2_ptr,
-                                           float *out_ptr) {
-  hipStream_t stream;
-  checkCUDA(get_legion_stream(&stream));
-
-  hipEvent_t t_start, t_end;
-  if (m->profiling) {
-    hipEventCreate(&t_start);
-    hipEventCreate(&t_end);
-    hipEventRecord(t_start, stream);
-  }
-  // print_tensor<float>(in1_ptr, in1_domain.get_volume(), "input1:");
-  // print_tensor<float>(in2_ptr, in2_domain.get_volume(), "input2:");
-  ElementBinary::forward_kernel(m, in1_ptr, in2_ptr, out_ptr, stream);
-  // print_tensor<float>(out_ptr, in1_domain.get_volume(), "output:");
-  if (m->profiling) {
-    hipEventRecord(t_end, stream);
-    checkCUDA(hipEventSynchronize(t_end));
-    float elapsed = 0;
-    checkCUDA(hipEventElapsedTime(&elapsed, t_start, t_end));
-    hipEventDestroy(t_start);
-    hipEventDestroy(t_end);
-    char const *opName;
-    switch (m->op_type) {
-      case OP_EW_ADD:
-        opName = "Add";
-        break;
-      case OP_EW_SUB:
-        opName = "Sub";
-        break;
-      case OP_EW_MUL:
-        opName = "Mul";
-        break;
-      case OP_EW_DIV:
-        opName = "Div";
-        break;
-      default:
-        assert(false);
-    }
-    log_measure.debug("[%s] forward time (CF) = %.2fms\n", opName, elapsed);
-  }
-}
-
-__global__ void elewise_binary_backward_kernel(coord_t volume,
-                                               float const alpha,
-                                               float const beta,
-                                               OperatorType type,
-                                               float const *out_grad,
-                                               float const *in1,
-                                               float const *in2,
-                                               float *in1_grad,
-                                               float *in2_grad) {
-  CUDA_KERNEL_LOOP(i, volume) {
-    switch (type) {
-      case OP_EW_ADD: {
-        in1_grad[i] = alpha * out_grad[i] + beta * in1_grad[i];
-        in2_grad[i] = alpha * out_grad[i] + beta * in2_grad[i];
-        break;
-      }
-      case OP_EW_SUB: {
-        in1_grad[i] = alpha * out_grad[i] + beta * in1_grad[i];
-        in2_grad[i] = -alpha * out_grad[i] + beta * in2_grad[i];
-        break;
-      }
-      case OP_EW_MUL: {
-        in1_grad[i] = alpha * out_grad[i] * in2[i] + beta * in1_grad[i];
-        in2_grad[i] = alpha * out_grad[i] * in1[i] + beta * in2_grad[i];
-        break;
-      }
-      case OP_EW_DIV: {
-        in1_grad[i] = alpha * out_grad[i] / in2[i] + beta * in1_grad[i];
-        in2_grad[i] = -alpha * out_grad[i] * in1[i] / (in2[i] * in2[i]) +
-                      beta * in2_grad[i];
-        break;
-      }
-      default:
-        assert(false);
-    }
-  }
-}
-
-/*static*/
-void ElementBinary::backward_kernel(ElementBinaryMeta const *m,
-                                    float const *out_grad_ptr,
-                                    float const *in1_ptr,
-                                    float const *in2_ptr,
-                                    float *in1_grad_ptr,
-                                    float *in2_grad_ptr,
-                                    hipStream_t stream) {
+void backward_kernel(ElementBinaryMeta const *m,
+                     float const *out_grad_ptr,
+                     float const *in1_ptr,
+                     float const *in2_ptr,
+                     float *in1_grad_ptr,
+                     float *in2_grad_ptr,
+                     hipStream_t stream) {
   checkCUDA(hipblasSetStream(m->handle.blas, stream));
   checkCUDNN(miopenSetStream(m->handle.dnn, stream));
 
@@ -341,61 +403,7 @@ void ElementBinary::backward_kernel(ElementBinaryMeta const *m,
   }
 }
 
-/*static*/
-void ElementBinary::backward_kernel_wrapper(ElementBinaryMeta const *m,
-                                            float const *out_grad_ptr,
-                                            float const *in1_ptr,
-                                            float const *in2_ptr,
-                                            float *in1_grad_ptr,
-                                            float *in2_grad_ptr) {
-  hipStream_t stream;
-  checkCUDA(get_legion_stream(&stream));
-  hipEvent_t t_start, t_end;
-  if (m->profiling) {
-    hipEventCreate(&t_start);
-    hipEventCreate(&t_end);
-    hipEventRecord(t_start, stream);
-  }
-
-  ElementBinary::backward_kernel(
-      m, out_grad_ptr, in1_ptr, in2_ptr, in1_grad_ptr, in2_grad_ptr, stream);
-  // elewise_binary_backward_kernel<<<GET_BLOCKS(out_grad_domain.get_volume()),
-  // CUDA_NUM_THREADS>>>( out_grad_domain.get_volume(), alpha, alpha,
-  // ele->op_type, out_grad_ptr, in1_ptr, in2_ptr, in1_grad_ptr, in2_grad_ptr);
-  if (m->profiling) {
-    hipEventRecord(t_end, stream);
-    checkCUDA(hipEventSynchronize(t_end));
-    float elapsed = 0;
-    checkCUDA(hipEventElapsedTime(&elapsed, t_start, t_end));
-    hipEventDestroy(t_start);
-    hipEventDestroy(t_end);
-    char const *opName;
-    switch (m->op_type) {
-      case OP_EW_ADD:
-        opName = "Add";
-        break;
-      case OP_EW_SUB:
-        opName = "Sub";
-        break;
-      case OP_EW_MUL:
-        opName = "Mul";
-        break;
-      case OP_EW_DIV:
-        opName = "Div";
-        break;
-      default:
-        assert(false);
-    }
-    printf("[%s] backward time (CB) = %.2fms\n", opName, elapsed);
-  }
-}
-
-ElementBinaryMeta::ElementBinaryMeta(FFHandler handler) : OpMeta(handler) {
-  checkCUDNN(miopenCreateTensorDescriptor(&input1Tensor));
-  checkCUDNN(miopenCreateTensorDescriptor(&input2Tensor));
-  checkCUDNN(miopenCreateTensorDescriptor(&outputTensor));
-  checkCUDNN(miopenCreateReduceTensorDescriptor(&reduceAddDesc));
-  op_type = OP_NOOP;
-}
-
+} // namespace Internal
+} // namespace ElementBinary
+} // namespace Kernels
 }; // namespace FlexFlow
