+++ src/runtime/simulator.cu
@@ -18,9 +18,9 @@
 #include "flexflow/ops/batch_norm.h"
 #include "flexflow/ops/concat.h"
 #include "flexflow/ops/conv_2d.h"
-#include "flexflow/ops/element_binary.h"
 #include "flexflow/ops/element_unary.h"
 #include "flexflow/ops/embedding.h"
+#include "flexflow/ops/kernels/element_binary_kernels.h"
 #include "flexflow/ops/linear.h"
 #include "flexflow/ops/pool_2d.h"
 #include "flexflow/ops/transpose.h"
